voice-to-markdown/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
websockets==14.1
pydantic==2.10.4
openai-whisper==20240930
requests==2.32.3
python-dotenv==1.0.1